ryu-line4/requirements.txt
line-bot-sdk
urllib3<2
requests
boto3
protobuf
openai==0.28.1